+++ src/AudienceSegments.php
@@ -23,6 +23,17 @@ class AudienceSegments
         ]);
     }
 
+    public function create(string $name, array $filter, ?string $description = null): mixed
+    {
+        return $this->client->query(method: 'POST', endpoint: 'v1/audience-segments', options: [
+            'json' => Util::omitNull([
+                'name' => $name,
+                'filter' => $filter,
+                'description' => $description,
+            ])
+        ]);
+    }
+
     public function get(string $audience_segment_id): mixed
     {
         return $this->client->query(method: 'GET', endpoint: 'v1/audience-segments/' . $audience_segment_id);